cleaned fsd
