+++ src/entities/user/index.ts
@@ -1,6 +1,7 @@
 export { getMe } from './model/api/user.service';
 export { useAuthMeQuery } from './model/hooks/useAuthMeQuery';
+export type { MeResponse, User, UserState } from './model/types/user.types';
+export { userRoleEnum } from './model/types/user-role.enum';
+export type { UserRole } from './model/types/user-role.types';
 export { selectUser } from './model/user.selectors';
 export { userStore, userStore as useUserStore } from './model/user.store';
-export type { MeResponse, User, UserRole, UserState } from './model/user.types';
-export { userRoleEnum } from './model/user-role.enum';
