+++ src/entities/user/model/types/user-role.enum.ts
@@ -1,6 +1,6 @@
 import z from 'zod';
 
-import { UserRole } from './user.types';
+import { UserRole } from './user-role.types';
 
 const userRoles = Object.values(UserRole) as UserRole[];
