+++ src/entities/user/model/types/user-role.types.ts
@@ -1,0 +1,6 @@
+export const UserRole = {
+  CANDIDATE: 'CANDIDATE',
+  RECRUITER: 'RECRUITER',
+} as const;
+
+export type UserRole = (typeof UserRole)[keyof typeof UserRole];
