+++ src/entities/user/model/user.selectors.ts
@@ -1,3 +1,3 @@
-import { UserState } from './user.types';
+import { UserState } from './types/user.types';
 
 export const selectUser = (state: UserState): UserState['user'] => state.user;
