+++ src/features/auth/index.ts
@@ -1,8 +1,12 @@
 export { useLogoutMutation } from '../../entities/session/model/hooks/useLogoutMutation';
-export type { AuthErrorResponse, AuthMethod, AuthMode } from './model';
 export { isAuthMode, isUserRole } from './model/guards';
 export { useAttachAuthToken } from './model/hooks/useAttachAuthToken';
 export { useLoginMutation } from './model/hooks/useLoginMutation';
 export { useRegisterMutation } from './model/hooks/useRegisterMutation';
+export type {
+  AuthErrorResponse,
+  AuthMethod,
+  AuthMode,
+} from './model/types/auth.types';
 export { AuthForm } from './ui/auth-form';
 export { RoleForm } from './ui/auth-role-form';
