--- src/features/auth/model/index.ts
@@ -1,11 +1,0 @@
-export type {
-  AuthShema,
-  EmailShema,
-  PhoneShema,
-} from './schema/AuthForm.shema';
-export { authSchema, emailSchema, phoneSchema } from './schema/AuthForm.shema';
-export type {
-  AuthErrorResponse,
-  AuthMethod,
-  AuthMode,
-} from './types/auth.types';
