+++ src/features/auth/ui/auth-role-form/components/selection-button/SelectionRoleButton.tsx
