+++ src/shared/ui/button/index.ts
@@ -1,1 +1,1 @@
-export { Button } from './button';
+export { Button } from './ui/button';
