+++ src/shared/ui/card/index.ts
@@ -1,1 +1,1 @@
-export { Card } from './card';
+export { Card } from './ui/card';
