+++ src/shared/ui/error-field/index.ts
@@ -1,1 +1,1 @@
-export { ErrorField } from './ErrorField';
+export { ErrorField } from './ui/ErrorField';
