+++ src/shared/ui/input/index.ts
@@ -1,1 +1,1 @@
-export { Input } from './input';
+export { Input } from './ui/input';
