+++ src/shared/ui/input/ui/input.stories.tsx
