+++ src/shared/ui/label/index.ts
@@ -1,1 +1,1 @@
-export { Label } from './label';
+export { Label } from './ui/label';
