+++ src/shared/ui/label/ui/label.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/nextjs-vite';
 
-import { Input } from '../input';
+import { Input } from '../../input';
 
 import { Label } from './label';
