+++ src/shared/ui/select/index.ts
@@ -1,1 +1,1 @@
-export { Select } from './Select';
+export { Select } from './ui/Select';
