+++ src/shared/ui/skeleton/index.ts
@@ -1,1 +1,1 @@
-export { Skeleton } from './skeleton';
+export { Skeleton } from './ui/skeleton';
