+++ src/widgets/auth-card/ui/components/auth-policy/AuthPolicy.tsx
