+++ src/widgets/auth-card/ui/components/auth-policy/index.ts
