+++ src/widgets/auth-card/ui/components/auth-switcher/AuthSwitcher.tsx
