+++ src/widgets/auth-card/ui/components/auth-switcher/index.ts
