+++ src/widgets/auth-role-card/ui/AuthRoleCard.tsx
