+++ src/widgets/auth-role-card/ui/navigate-buttons/NavigateButtons.tsx
