+++ src/widgets/auth-role-card/ui/navigate-buttons/index.ts
